requirements.txt
kafka-python
grpcio
grpcio-tools